+++ Win32Frame/Win32Frame/Object.h
@@ -10,7 +10,7 @@ public:
 	Object();
 	virtual ~Object();
 
-	virtual void Update( float dt );
+	virtual void Update( double dt );
 	virtual void Draw( HDC hdc ) = 0;
 
 	virtual void Translate( float x, float y );
